requirements.txt
streamlit>=1.24.0
fpdf>=1.7.2
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.1
groq